Ensure the library commentary is not empty
Previously the content of "Commentary" was a single empty line.
Make "Overview" and "Entry points" sub-sections, instead of top-level
sections.

Meanwhile, keep "INSTALL" and "TODO" as top-level sections.  Otherwise
they would appear in the documentation shown by Melpa, package.el and
other tools, and that wouldn't be desirable.  The "INSTALL" section is
not relevant when using package.el to install.