Store input buffer position in a buffer-local variable instead of in connection.
A process sentinel only has access to the process and the associated event (input bytes for our
connection to PostgreSQL); it doesn't have access to our pgcon connection structure. With this
change, a process sentinel can access the current input position in the process-buffer by reading
the buffer-local pgcon-position variable. This should allow us to handle asynchronous inputs from
PostgreSQL, such as NoticeMessages.