Modifications to initialization for pg-geometry and pg-gis libraries.
If you require the pg-geometry library after you establish a PostgreSQL connection, you will need to
call the new pg-geometry-setup function in order for the type serialization and deserialization
support to be set up. This will look up the OID for the different geometry types and hook them into
the parsing machinery. Likewise for pg-gis, you need to call the (pre-existing) pg-postgis-setup
function, which also loads the PostGIS extension if necessary.