Fix timezone handling
Timezone information would be lost during parsing (passing a "+04"
suffix directly into encode-time does not work, since either an offset
in seconds or a named TZ is expected). Thus we replace manual parsing
by parse-iso8601-time-string.

Furthermore we widen the regexp to allow "+04:00" as well as "Z" or
"z" timezone suffixes, which are perfectly fine in ISO8601.

The timezone would get lost during serialization, too. We add it to
the respective format string.