New function pg-connect-plist to replace pg-connect.
The new function pg-connect-plist is similar to pg-connect, but takes keyword arguments instead of
optional arguments, to make it extensible in the future. The function pg-connect is deprecated.
Likewise, a new macro with-pg-connection-plist should be used instead of with-pg-connection.

The recently introduced function pg-connect/direct-tls has been deprecated; use the :direct-tls
option to pg-connect-plist instead.