Support for distinct value for SQL NULLs
The pg-null-marker variable can be set to a value such as "ω" which represents
SQL NULL values. The current default setting is nil, which is the value which
was used to represent NULL up to now.

Note that NULL in JSON and JSONB values remains represented by the special value :null.