Recognize type alises in pg-exec-prepared
Type aliases are recognized as type names in `pg-exec-prepared`, in additional
to the type names present in the `pg_type` system table. This means that you can
use `bigint` as an alternative to `int8`, `real` as an alternative to `float4`,
`bit varying` as an alternative to `bit`, for example. The aliases are also the
canonical type names as shown by function `pg_typeof`.